placed personViewSlim styles in a new file
