+++ frontend/app/src/people/auth/authQR.tsx
@@ -45,7 +45,7 @@ export default function AuthQR(props: any) {
         if (i > 100) {
           if (interval) clearInterval(interval);
         }
-      } catch (e) {}
+      } catch (e) { }
     }, 3000);
   }
   async function getChallenge() {
